IN_DEVELOPMENT - bug 19: Track Created, Updated, and Deleted Taxa http://129.67.24.160/bugzilla/show_bug.cgi?id=19
Added (some) triggers to the mysql test database schema, also auto_incremented the primary key for Plant_name. Added beginning of the DBUnit tests, but not straightforward to use these as we cannot easily tear down the database (e.g. delete Plant_name) and insert fresh data into UAT, so we're reduced to inserting a known dataset and removing it at the end of each test so we have a known state.