Build docker images for harvester, portal, apache
Integrate docker into the maven build. Tomcat images for the portal and
harvester are built with the compiled wars and apache is build with
static assets.

Application configuration was pulled in to codebase and centralized in
application.properties and test.properties for testing overrides.

All the various configuration variants via maven profiles were axed as
they are no longer needed