Adds a local env file and instructions to load images locally
In order for images to be loaded by POWO, the application property
cdn.key must be set. This is a secret, so should be provided via
a CDN_KEY environment variable. This variable will be translated by
spring into the cdn.key application property.

When developing using docker-compose, environment variables are loaded
from an optional .env file. This commit adds instructions on how to
use this file, and instructs docker-compose to load the CDN_KEY
variable for the portal and harvester modules.