Adds automatic compilation and update of portal JS and CSS
The first step of adding this feature was to make sure the gulp
watch task worked. As node 14 doesn't support gulp v3, this meant
either using node 10 locally, or updating the project to use node 14.
I chose the latter, to keep the code up to date with latest software
versions. Once this was done I updated gulp from version 3 to 4.

When gulp recompiles the files, the new versions need to be picked
up by the ingress service. This commit achieves this by linking the
powo-portal dist directory to the ingress container's /www/data
directory.