Adds browsersync for live reload and asset timestamps for cache busting
This commit updates the watch task to use the browser-sync library to
reload the browser when the JS or CSS changes. The browsersync server
proxies requests to the backend ingress server (port 10080).

The AssetHelper is updated so that in development mode it appends a
timestamp to the asset path. This prevents the browser using cached
frontend assets.