Adds a test for AssetHelper in development mode
The test checks that the compiled asset paths have a timestamp
query parameter appended to the end (?t=...). The change to the
use of the helper, to use 3 brackets ({{{...}}}) is to prevent
handlebars encoding the query string. It is safe to use an
unencoded value here, as we are only using the helper for file
paths, not user input.

In order to test the injected developmentMode value, the test
is run using the Spring test runner. The AssetHelper is added
to the test application context so that Spring can be used
to provide the value for developmentMode.