Fixes a bug where location suggestions werent being returned on POWO
This was because the cfq query "Plantae" was being applied to all the
suggesters, causing the location suggester to return nothing. This has
been "fixed" by changing the type of the suggester from
AnalysingInfixLookup to AnalysingLookup. Further work is needed to make
infix searchers work for the location suggester (which was never working
it seems).