remove openlayers and update custom build docs
We need to do this because the current version of openlayers does NOT build due to changes to how our firewall manages downloading unsafe files (e.g. zip files).

The new version of openlayers is called `ol` and does not have the same setup for how to produce a custom build. Instead it uses webpack/rollup but we do not currently use these, so we are sticking with the vendored libraries we have already built.