remove annotation creation logic in harvester
After a week of debugging an issue relating to annotations saving incorrectly, this is an attempt to side step the problem by just removing them.

This may seem controversial but it is not clear that annotations actually do anything other than provide a log of what is inserted. Because we always build from a fresh database, and processor logic performs updates if necessary, this means that the annotation saving is redundant. And currently we save ~9million records every time we harvest.

I'm hoping this will work out well in the end.